nlp/requirements.txt
numpy
Jinja2
pandas
Flask
flask-cors
pyarrow
werkzeug
python-dotenv
openai
torch
transformers
paramiko
scikit-learn
fast-pytorch-kmeans
tiktoken
tenacity
SQLAlchemy
psycopg2-binary
requests
